feat: Migrate gas limit to u64 (#1538)
## What ❔

⚠️ **Warning: this PR requires DB migration** ⚠️ 

In preparation for the 1.5.0 upgrade:

- We now use u64 to represent the `gasLimit`. In most places we already
have bigdecimal / U256 there, so there are not many changes there.
- Since the gas limit will be able to exceed u32 in 1.5.0, it means that
the `block.gaslimit` should be able to exceed it to in the 1.5.0 upgrade
too. Our values should return consistent block gas limit with the system
contracts, so we have to return it here also. Due to this part, we will
require a migration, to store the gas limit for each block.

Follow-up tasks needed:
- Add a seal criteria to prevent the sum of gas limits of transactions
in a batch exceed the batch gas limit. Note, that in the first release
of the 1.5.0 this property won't be enforced.

## Why ❔

To prepare for 1.5.0 upgrade, where gas limit can exceed u32::MAX

## Checklist

<!-- Check your PR fulfills the following items. -->
<!-- For draft PRs check the boxes as you complete them. -->

- [ ] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [ ] Tests for the changes have been added / updated.
- [ ] Documentation comments have been added / updated.
- [ ] Code has been formatted via `zk fmt` and `zk lint`.
- [ ] Spellcheck has been run via `zk spellcheck`.
- [ ] Linkcheck has been run via `zk linkcheck`.

+++ core/lib/config/src/configs/chain.rs
@@ -104,7 +104,7 @@ pub struct StateKeeperConfig {
     /// The max number of gas to spend on an L1 tx before its batch should be sealed by the gas sealer.
     pub max_single_tx_gas: u32,
 
-    pub max_allowed_l2_tx_gas_limit: u32,
+    pub max_allowed_l2_tx_gas_limit: u64,
 
     /// Configuration option for tx to be rejected in case
     /// it takes more percentage of the block capacity than this value.
